REFACTOR: move base token case to high level
